Move surface and transform to their own directories
