+++ src_c/surface/surface.c
@@ -29,7 +29,7 @@
 
 #include "structmember.h"
 #include "pgcompat.h"
-#include "doc/surface_doc.h"
+#include "surface_doc.h"
 
 /* stdint.h is missing from some versions of MSVC. */
 #ifdef _MSC_VER
